Refactored types
Moved types into their own files within the types directory
