+++ src/php/mod.rs
@@ -4,4 +4,3 @@ pub mod execution_data;
 pub mod function;
 pub mod module;
 pub mod types;
-pub mod zval;
