+++ src/php/types/mod.rs
@@ -1,12 +1,13 @@
+pub mod array;
+pub mod long;
+pub mod string;
+pub mod zval;
+
 use std::{ffi::c_void, ptr};
 
-use crate::{
-    bindings::{
-        zend_long, zend_string, zend_strpprintf, zend_type, GC_FLAGS_MASK, GC_FLAGS_SHIFT,
-        GC_INFO_SHIFT, IS_MIXED, IS_STR_INTERNED, MAY_BE_ANY, MAY_BE_BOOL, _IS_BOOL,
-        _ZEND_IS_VARIADIC_BIT, _ZEND_SEND_MODE_SHIFT, _ZEND_TYPE_NULLABLE_BIT,
-    },
-    functions::c_str,
+use crate::bindings::{
+    zend_type, IS_MIXED, MAY_BE_ANY, MAY_BE_BOOL, _IS_BOOL, _ZEND_IS_VARIADIC_BIT,
+    _ZEND_SEND_MODE_SHIFT, _ZEND_TYPE_NULLABLE_BIT,
 };
 
 use super::enums::DataType;
@@ -84,42 +85,5 @@ impl ZendType {
         } else {
             0
         }) | Self::arg_info_flags(pass_by_ref, is_variadic)
-    }
-}
-
-/// Internal identifier used for a long.
-/// The size depends on the system architecture. On 32-bit systems,
-/// a ZendLong is 32-bits, while on a 64-bit system it is 64-bits.
-pub type ZendLong = zend_long;
-
-/// String type used in the Zend internals.
-/// The actual size of the 'string' differs, as the
-/// end of this struct is only 1 char long, but the length
-/// inside the struct defines how many characters are in the string.
-pub type ZendString = zend_string;
-
-impl ZendString {
-    /// Creates a new Zend string.
-    ///
-    /// Note that this returns a raw pointer, and will not be freed by
-    /// Rust.
-    ///
-    /// # Parameters
-    ///
-    /// * `str_` - The string to create a Zend string from.
-    pub fn new<S>(str_: S) -> *mut Self
-    where
-        S: AsRef<str>,
-    {
-        let str_ = str_.as_ref();
-        unsafe { zend_strpprintf(str_.len() as u64, c_str(str_)) }
-    }
-
-    /// Translation of the `ZSTR_IS_INTERNED` macro.
-    /// zend_string.h:76
-    pub(crate) unsafe fn is_interned(&self) -> bool {
-        (((self.gc.u.type_info >> GC_INFO_SHIFT) & (GC_FLAGS_MASK >> GC_FLAGS_SHIFT))
-            & IS_STR_INTERNED)
-            != 0
     }
 }
